Migrated Resilience component away from hooks component to middleware component
