+++ src/resilience/_module.ts
@@ -1,2 +1,4 @@
-export * from "@/resilience/middlewares/_module.js";
+export * from "@/resilience/fallback/fallback.js";
+export * from "@/resilience/retry/retry.js";
+export * from "@/resilience/timeout/timeout.js";
 export * from "@/resilience/resilience.errors.js";
