+++ src/resilience/fallback/_module.ts
@@ -1,0 +1,1 @@
+export * from "@/resilience/fallback/fallback.js";
