+++ src/resilience/fallback/fallback.test.ts
@@ -1,0 +1,5 @@
+import { describe, test } from "vitest";
+
+describe("function: fallback", () => {
+    test.todo("Write tests!!!");
+});
