--- src/resilience/middlewares/_module.ts
@@ -1,5 +1,0 @@
-export * from "@/resilience/middlewares/dynamic/_module.js";
-export * from "@/resilience/middlewares/fallback/_module.js";
-export * from "@/resilience/middlewares/observe/_module.js";
-export * from "@/resilience/middlewares/retry/_module.js";
-export * from "@/resilience/middlewares/timeout/_module.js";
