--- src/resilience/middlewares/dynamic/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/middlewares/dynamic/dynamic.middleware.js";
