--- src/resilience/middlewares/fallback/_module.ts
@@ -1,2 +1,0 @@
-export * from "@/resilience/middlewares/fallback/fallback.middleware.js";
-export type * from "@/resilience/middlewares/fallback/fallback.types.js";
