--- src/resilience/middlewares/observe/_module.ts
@@ -1,2 +1,0 @@
-export * from "@/resilience/middlewares/observe/observe.middleware.js";
-export type * from "@/resilience/middlewares/observe/observe.types.js";
