--- src/resilience/middlewares/retry/_module.ts
@@ -1,2 +1,0 @@
-export * from "@/resilience/middlewares/retry/retry.middleware.js";
-export type * from "@/resilience/middlewares/retry/retry.types.js";
