--- src/resilience/middlewares/timeout/_module.ts
@@ -1,2 +1,0 @@
-export * from "@/resilience/middlewares/timeout/timeout.middleware.js";
-export type * from "@/resilience/middlewares/timeout/timeout.type.js";
