+++ src/resilience/retry/retry.test.ts
@@ -1,0 +1,5 @@
+import { describe, test } from "vitest";
+
+describe("function: retry", () => {
+    test.todo("Write tests!!!");
+});
