+++ src/resilience/timeout/timeout.test.ts
@@ -1,0 +1,5 @@
+import { describe, test } from "vitest";
+
+describe("function: timeout", () => {
+    test.todo("Write tests!!!");
+});
