--- src/resilience/utilities/abort-and-fail/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/utilities/abort-and-fail/abort-and-fail.js";
