--- src/resilience/utilities/timeout-and-fail/_module.ts
@@ -1,1 +1,0 @@
-export * from "@/resilience/utilities/timeout-and-fail/timeout-and-fail.js";
